requirements.txt
pymongo
fastapi
uvicorn
pydantic
python-dotenv
aiohttp
redis
PyJWT
bcrypt
email-validator
uvloop
google-genai
boto3
